Switch to own shared lib :(
