+++ api/package.json
@@ -15,13 +15,12 @@
     "toc": "doctoc --github --maxlevel 2 README.md"
   },
   "dependencies": {
-    "@shared-libs/message-utils": "file:../shared-libs/message-utils",
+    "@shared-libs/registration-utils": "file:../shared-libs/registration-utils",
     "@shared-libs/server-checks": "file:../shared-libs/server-checks",
     "@shared-libs/tombstone-utils": "file:../shared-libs/tombstone-utils",
     "@shared-libs/view-map-utils": "file:../shared-libs/view-map-utils",
     "async": "^2.6.0",
     "body-parser": "^1.18.2",
-    "bikram-sambat-bootstrap": "^1.4.2",
     "buffer-shims": "^1.0.0",
     "bulk-docs-utils": "file:../shared-libs/bulk-docs-utils",
     "compression": "^1.7.2",
@@ -29,7 +28,6 @@
     "express": "^4.16.3",
     "fast-csv": "^2.4.1",
     "google-libphonenumber": "^3.1.8",
-    "gsm": "^0.1.4",
     "helmet": "^3.12.1",
     "http-proxy": "^1.17.0",
     "lineage": "file:../shared-libs/lineage",
@@ -57,13 +55,12 @@
     "doctoc": "^1.3.1"
   },
   "bundledDependencies": [
-    "@shared-libs/message-utils",
+    "@shared-libs/registration-utils",
     "@shared-libs/server-checks",
     "@shared-libs/tombstone-utils",
     "@shared-libs/view-map-utils",
     "async",
     "body-parser",
-    "bikram-sambat-bootstrap",
     "buffer-shims",
     "bulk-docs-utils",
     "compression",
@@ -71,7 +68,6 @@
     "express",
     "fast-csv",
     "google-libphonenumber",
-    "gsm",
     "helmet",
     "http-proxy",
     "lineage",
