+++ sentinel/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@shared-libs/message-utils": "file:../shared-libs/message-utils",
+    "@shared-libs/registration-utils": "file:../shared-libs/registration-utils",
     "@shared-libs/server-checks": "file:../shared-libs/server-checks",
     "@shared-libs/tombstone-utils": "file:../shared-libs/tombstone-utils",
     "async": "^2.6.0",
@@ -38,6 +39,7 @@
   },
   "bundledDependencies": [
     "@shared-libs/message-utils",
+    "@shared-libs/registration-utils",
     "@shared-libs/tombstone-utils",
     "async",
     "bikram-sambat",
