+++ sentinel/tests/unit/utils.js
@@ -3,7 +3,7 @@ var db = require('../../src/db-nano'),
     assert = require('chai').assert,
     utils = require('../../src/lib/utils'),
     config = require('../../src/config'),
-    messageUtils = require('@shared-libs/message-utils');
+    registrationUtils = require('@shared-libs/registration-utils');
 
 describe('utils', () => {
     afterEach(() => sinon.restore());
@@ -205,7 +205,7 @@ describe('utils', () => {
     });
 
     it('getRegistrations queries by id if given', () => {
-        sinon.stub(messageUtils, 'isValidRegistration').returns(true);
+        sinon.stub(registrationUtils, 'isValidRegistration').returns(true);
         sinon.stub(config, 'getAll').returns({ config: 'all' });
         const expectedDoc = { _id: 'a' };
         const expected = [ { doc: expectedDoc } ];
@@ -215,8 +215,8 @@ describe('utils', () => {
             assert.equal(err, null);
             assert.deepEqual(actual, [ expectedDoc ]);
             assert.equal(view.callCount, 1);
-            assert.equal(messageUtils.isValidRegistration.callCount, 1);
-            assert.deepEqual(messageUtils.isValidRegistration.args[0], [expectedDoc, { config: 'all' }]);
+            assert.equal(registrationUtils.isValidRegistration.callCount, 1);
+            assert.deepEqual(registrationUtils.isValidRegistration.args[0], [expectedDoc, { config: 'all' }]);
             assert.equal(view.args[0][0], 'medic-client');
             assert.equal(view.args[0][1], 'registered_patients');
             assert.equal(view.args[0][2].key, given);
@@ -225,7 +225,7 @@ describe('utils', () => {
     });
 
     it('getRegistrations queries by ids if given', () => {
-        sinon.stub(messageUtils, 'isValidRegistration').returns(true);
+        sinon.stub(registrationUtils, 'isValidRegistration').returns(true);
         sinon.stub(config, 'getAll').returns({ config: 'all' });
         const expectedDoc1 = { id: 'a' };
         const expectedDoc2 = { id: 'b' };
@@ -234,9 +234,9 @@ describe('utils', () => {
         const view = sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: expected });
         return utils.getRegistrations({ db: db, ids: given }, (err, actual) => {
             assert.equal(err, null);
-            assert.equal(messageUtils.isValidRegistration.callCount, 2);
-            assert.deepEqual(messageUtils.isValidRegistration.args[0], [expectedDoc1, { config: 'all' }]);
-            assert.deepEqual(messageUtils.isValidRegistration.args[1], [expectedDoc2, { config: 'all' }]);
+            assert.equal(registrationUtils.isValidRegistration.callCount, 2);
+            assert.deepEqual(registrationUtils.isValidRegistration.args[0], [expectedDoc1, { config: 'all' }]);
+            assert.deepEqual(registrationUtils.isValidRegistration.args[1], [expectedDoc2, { config: 'all' }]);
             assert.deepEqual(actual, [expectedDoc1, expectedDoc2 ]);
             assert.equal(view.callCount, 1);
             assert.equal(view.args[0][0], 'medic-client');
@@ -256,12 +256,12 @@ describe('utils', () => {
     });
 
     it('getRegistrations only returns valid registrations', () => {
-      sinon.stub(messageUtils, 'isValidRegistration');
+      sinon.stub(registrationUtils, 'isValidRegistration');
       sinon.stub(config, 'getAll').returns({ config: 'all' });
       const docs = [{ _id: 'a' }, { _id: 'b' }, { _id: 'c' }, { _id: 'd' }, { _id: 'e' }, { _id: 'f' }];
       sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: docs.map(doc => ({ doc: doc })) });
 
-      messageUtils.isValidRegistration
+      registrationUtils.isValidRegistration
         .withArgs({ _id: 'a' }).returns(true)
         .withArgs({ _id: 'b' }).returns(false)
         .withArgs({ _id: 'c' }).returns(false)
@@ -271,7 +271,7 @@ describe('utils', () => {
 
       return utils.getRegistrations({ db: db, ids: ['111', '222'] }, (err, actual) => {
         assert.equal(err, null);
-        assert.equal(messageUtils.isValidRegistration.callCount, 6);
+        assert.equal(registrationUtils.isValidRegistration.callCount, 6);
         assert.deepEqual(actual, [{ _id: 'a' }, { _id: 'd' }]);
       });
     });
