+++ shared-libs/registration-utils/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@shared-libs/registration-utils",
+  "version": "1.0.0",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "mocha ./test"
+  },
+  "private": true,
+  "devDependencies": {
+    "chai": "^4.1.2",
+    "mocha": "^5.0.0",
+    "sinon": "4.0.1"
+  }
+}
